pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "luna-mcp-server"
version = "0.1.2"
description = "Indie Hub: Luna MCP Server bridging Puch MCP to Luna Services with developer tooling."
authors = [{ name = "Indie Hub / Mantej Singh" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi==0.110.*",
    "uvicorn[standard]",
    "python-dotenv",
    "pydantic>=2",
    "httpx",
    "PyGithub",
    "supabase",
    "Pillow",
    "python-multipart"
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_bearer_token", "tools", "github_oauth"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "mypy",
    "ruff",
    "pytest-cov"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
addopts = "-q --cov=mcp-bearer-token --cov=tools --cov-report=term-missing --cov-report=xml"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
